+++ packages/logic-core/src/data/solver/backtrack/backtrackSolver.ts
@@ -1,5 +1,3 @@
-import { EventIterator } from 'event-iterator';
-import GridData from '../../grid.js';
 import { instance as banPatternInstance } from '../../rules/banPatternRule.js';
 import { instance as cellCountInstance } from '../../rules/cellCountRule.js';
 import { instance as regionAreaInstance } from '../../rules/regionAreaRule.js';
@@ -7,7 +5,6 @@ import { instance as sameShapeInstance } from '../../rules/sameShapeRule.js';
 import { instance as symbolsPerRegionInstance } from '../../rules/symbolsPerRegionRule.js';
 import { instance as undercluedInstance } from '../../rules/undercluedRule.js';
 import { instance as uniqueShapeInstance } from '../../rules/uniqueShapeRule.js';
-import { Serializer } from '../../serializer/allSerializers.js';
 import { instance as areaNumberInstance } from '../../symbols/areaNumberSymbol.js';
 import { instance as dartInstance } from '../../symbols/dartSymbol.js';
 import { instance as galaxyInstance } from '../../symbols/galaxySymbol.js';
@@ -16,10 +13,10 @@ import { instance as lotusInstance } from '../../symbols/lotusSymbol.js';
 import { instance as minesweeperInstance } from '../../symbols/minesweeperSymbol.js';
 import { instance as myopiaInstance } from '../../symbols/myopiaSymbol.js';
 import { instance as viewpointInstance } from '../../symbols/viewpointSymbol.js';
-import Solver from '../solver.js';
 import { instance as connectAllInstance } from '../z3/modules/connectAllModule.js';
+import EventIteratingSolver from '../eventIteratingSolver.js';
 
-export default class BacktrackSolver extends Solver {
+export default class BacktrackSolver extends EventIteratingSolver {
   private static readonly supportedInstrs = [
     areaNumberInstance.id,
     viewpointInstance.id,
@@ -44,40 +41,10 @@ export default class BacktrackSolver extends Solver {
   public readonly description =
     'Solves puzzles using backtracking with optimizations (blazingly fast). Support most rules and symbols (including underclued).';
 
-  public async *solve(grid: GridData): AsyncGenerator<GridData | null> {
-    const worker = new Worker(
-      new URL(`./backtrackWorker.js`, import.meta.url),
-      {
-        type: 'module',
-      }
-    );
-
-    try {
-      const iterator = new EventIterator<GridData>(({ push, stop, fail }) => {
-        worker.postMessage(Serializer.stringifyGrid(grid.resetTiles()));
-
-        worker.addEventListener('message', (e: MessageEvent<string | null>) => {
-          if (e.data) {
-            push(Serializer.parseGrid(e.data));
-          } else {
-            stop();
-          }
-        });
-
-        worker.addEventListener('error', (e: ErrorEvent) => {
-          alert(`Error while solving!\n${e.message}`);
-          fail(e as unknown as Error);
-        });
-      });
-
-      for await (const solution of iterator) {
-        yield solution;
-      }
-
-      yield null;
-    } finally {
-      worker.terminate();
-    }
+  protected createWorker(): Worker {
+    return new Worker(new URL(`./backtrackWorker.js`, import.meta.url), {
+      type: 'module',
+    });
   }
 
   public isInstructionSupported(instructionId: string): boolean {