+++ packages/logic-core/src/data/solver/backtrack/backtrackWorker.ts
@@ -105,9 +105,10 @@ function translateToBTGridData(grid: GridData): BTGridData {
         continue;
       }
 
-      if (!module) throw new Error('Symbol not supported.');
+      if (!module && symbol.necessaryForCompletion)
+        throw new Error('Symbol not supported.');
 
-      modules.push(module);
+      if (module) modules.push(module);
     }
   }
 
@@ -316,7 +317,9 @@ function solveUnderclued(input: GridData): GridData | null {
 
     // console.log(`Trying (${x}, ${y}) with ${color}`);
 
-    const newGrid = grid.setTile(x, y, tile.withColor(color));
+    const newGrid = grid.fastCopyWith({
+      tiles: grid.setTile(x, y, tile.withColor(color)),
+    });
 
     // Solve
     let solution: GridData | undefined;
@@ -355,9 +358,13 @@ function solveUnderclued(input: GridData): GridData | null {
       if (!darkPossible && !lightPossible) return null;
 
       if (darkPossible && !lightPossible)
-        grid = grid.setTile(x, y, tile.withColor(Color.Dark));
+        grid = grid.fastCopyWith({
+          tiles: grid.setTile(x, y, tile.withColor(Color.Dark)),
+        });
       if (!darkPossible && lightPossible)
-        grid = grid.setTile(x, y, tile.withColor(Color.Light));
+        grid = grid.fastCopyWith({
+          tiles: grid.setTile(x, y, tile.withColor(Color.Light)),
+        });
     }
   }
 
@@ -396,6 +403,4 @@ onmessage = e => {
   postMessage(null);
 };
 
-// make typescript happy
-// eslint-disable-next-line import/no-anonymous-default-export
-export default null;
+export {};